refactor: split the data read-store god adapter along the service seam
postgres_data_read_store.py was a 665-line adapter carrying both the
streaming read path and the catalog/manifest/record-by-id reads behind one
DataReadStore port. Split into DataTableReadStore (PostgresTableReadStore:
streaming, filter SQL, keyset) and DataCatalogReadStore
(PostgresCatalogReadStore: catalog, manifest, record lookup), matching the
DataQueryService / DataCatalogService split. Schema→feature-table resolution
shared by both adapters lives in schema_features.py.
