+++ server/osa/domain/data/util/di/provider.py
@@ -3,7 +3,10 @@
 from dishka import provide
 
 from osa.config import Config
-from osa.domain.data.port.data_read_store import DataReadStore
+from osa.domain.data.port.data_read_store import (
+    DataCatalogReadStore,
+    DataTableReadStore,
+)
 from osa.domain.data.query.catalog import (
     GetDataRecordHandler,
     GetNodeCatalogHandler,
@@ -21,11 +24,13 @@ from osa.util.di.scope import Scope
 
 class DataProvider(Provider):
     @provide(scope=Scope.UOW)
-    def get_data_query_service(self, read_store: DataReadStore, config: Config) -> DataQueryService:
+    def get_data_query_service(
+        self, read_store: DataTableReadStore, config: Config
+    ) -> DataQueryService:
         return DataQueryService(read_store=read_store, config=config)
 
     @provide(scope=Scope.UOW)
-    def get_data_catalog_service(self, read_store: DataReadStore) -> DataCatalogService:
+    def get_data_catalog_service(self, read_store: DataCatalogReadStore) -> DataCatalogService:
         return DataCatalogService(read_store=read_store)
 
     read_records_table_handler = provide(ReadRecordsTableHandler, scope=Scope.UOW)
