+++ server/tests/unit/infrastructure/data/test_cursor_validation.py
@@ -27,8 +27,8 @@ from osa.domain.data.model.query_plan import (
 )
 from osa.domain.shared.error import ValidationError
 from osa.domain.shared.model.hook import ColumnDef
-from osa.domain.shared.model.srn import Domain, SchemaId
-from osa.infrastructure.data.postgres_data_read_store import PostgresDataReadStore
+from osa.domain.shared.model.srn import SchemaId
+from osa.infrastructure.data.postgres_table_read_store import PostgresTableReadStore
 from osa.infrastructure.persistence.feature_table import (
     FeatureSchema,
     build_feature_table,
@@ -38,10 +38,10 @@ from osa.infrastructure.persistence.tables import records_table
 SCHEMA = SchemaId.parse("compound@1.0.0")
 
 
-def _store() -> PostgresDataReadStore:
+def _store() -> PostgresTableReadStore:
     # The sort helpers never touch self.session; decoding happens before any DB
     # access, so a placeholder session is sufficient for these unit tests.
-    return PostgresDataReadStore(None, Domain("localhost"))  # type: ignore[arg-type]
+    return PostgresTableReadStore(None)  # type: ignore[arg-type]
 
 
 def _records_plan(cursor: str) -> QueryPlan:
@@ -84,13 +84,13 @@ class TestRecordsCursorValidation:
         # A FieldType.DATE metadata column is a sa.Date in the dynamic table;
         # the cursor carries its value as an ISO string — asyncpg rejects a
         # str bound against DATE, so the decoder must coerce by column type.
-        value = PostgresDataReadStore._coerce_cursor_value(
+        value = PostgresTableReadStore._coerce_cursor_value(
             "2026-01-02", sa.Column("assay_date", sa.Date())
         )
         assert isinstance(value, date)
 
     def test_datetime_metadata_cursor_value_coerced_to_datetime(self) -> None:
-        value = PostgresDataReadStore._coerce_cursor_value(
+        value = PostgresTableReadStore._coerce_cursor_value(
             "2026-01-02T03:04:05+00:00", sa.Column("measured_at", sa.DateTime(timezone=True))
         )
         assert isinstance(value, datetime)
@@ -115,7 +115,7 @@ class TestFeatureCursorValidation:
         # created_at is an implicit feature column; the type-driven coercion
         # must turn its ISO string back into a datetime before binding.
         ft = _feature_table()
-        value = PostgresDataReadStore._coerce_cursor_value(
+        value = PostgresTableReadStore._coerce_cursor_value(
             "2026-01-02T03:04:05+00:00", ft.c.created_at
         )
         assert isinstance(value, datetime)
@@ -124,4 +124,4 @@ class TestFeatureCursorValidation:
         # Sorting by id, the cursor sort value binds against BIGINT — a str
         # is coerced (and garbage raises ValueError → 400 via the sort path).
         ft = _feature_table()
-        assert PostgresDataReadStore._coerce_cursor_value("7", ft.c.id) == 7
+        assert PostgresTableReadStore._coerce_cursor_value("7", ft.c.id) == 7
